Remove asCommutativeRing from Heyting
